Fix #3099 resolving of configuration references. Fix getting template's default value.
